workers/ai_worker/pyproject.toml
[project]
name = "ai-tutor-worker"
version = "0.1.0"
description = "Internal AI/CV worker service for AI Tutor"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pillow>=10.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
cv = [
    "opencv-python>=4.10.0",
    "numpy>=1.26.0",
]
ocr = [
    "pytesseract>=0.3.13",
]
[tool.ruff]
line-length = 100
target-version = "py311"
